requirements.txt
ultralytics>=8.3.0 # YOLO architecture and tracker basics
torch>=2.0.0
torchvision
numpy>=1.22.0
opencv-python>=4.8.0
pyyaml>=6.0
pandas>=2.0.0
tqdm>=4.65.0
lapx>=0.5.4 # required for ByteTrack
